% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reading.R
\name{book_finish_ratios}
\alias{book_finish_ratios}
\title{find the number of finished and unfinished book reads over sequence of dates}
\usage{
book_finish_ratios(
  start_date = (Sys.Date() - 7),
  end_date = Sys.Date(),
  my_notes = NULL
)
}
\arguments{
\item{start_date}{a date or a string that can be parsed by lubridate's
`dmy()`.  Default Sys.Date() - 7.}

\item{end_date}{a date or a string that can be parsed by lubridate's
`dmy()`.  Default Sys.Date().}

\item{my_notes}{a book notes dataframe; default NULL}
}
\description{
Based on the start and finish dates in a book notes data frame, returns a
new data frame showing the number of books finished by each date and the
number of books still mid-read on each date.
}
\keyword{book}
\keyword{notes,}
\keyword{summary}
