% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reading.R
\name{get_page_count}
\alias{get_page_count}
\title{get the page count for a book from notes on it}
\usage{
get_page_count(source_note = NULL)
}
\arguments{
\item{one}{at a time for inputs}
}
\value{
describe the function's output
}
\description{
ZZZ block describe function
}
\details{
ZZZ big block describes function
}
\keyword{book}
\keyword{count}
\keyword{notes,}
\keyword{page}
