% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nehnar-package.R
\docType{package}
\name{nehnar-package}
\alias{nehnar}
\alias{nehnar-package}
\title{nehnar: a Bag of Functions We Use}
\description{
These functions are used by us for a few different
    things.  There are functions related to doing maps of places we've been
    and some that are related to keeping track of my reading.
}
\author{
\strong{Maintainer}: Adam George Dunn \email{agd1983@gmail.com}

}
\keyword{internal}
