% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reading.R
\name{read_book_notes}
\alias{read_book_notes}
\title{Read in My Book Notes}
\usage{
read_book_notes(source_file = NULL, diagnose_speed = FALSE)
}
\arguments{
\item{source_file}{path to the plain-text file to pull book notes from.
Will only work if used on a file with the same layout as mine.
Default NULL.}

\item{diagnose_speed}{a boolean; if TRUE, prints the time taken by the core
(slow) loop in the function.  Default FALSE.}
}
\value{
a data frame with bibliographic details of books as separate
  columns.  Also includes start and end dates of reading each book and
  keywords I've noted for each.
}
\description{
Reads in a plain text file of notes on books and returns a tidy data frame.
}
\details{
Designed for my own format of notes on books, including when I read them.
Won't work on anything that doesn't match that formatting.  Creates
tolerably pretty variables.
}
\keyword{data}
\keyword{idiosyncratic}
\keyword{munging,}
