% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reading.R
\name{check_book_progress}
\alias{check_book_progress}
\title{Show Book Progress on a Particular Day}
\usage{
check_book_progress(my_data, a_date, keep_all = FALSE)
}
\arguments{
\item{my_data}{a dataframe of book notes.  Must include start and finish
dates for reading attempts.  No default.}

\item{a_date}{a date or character object of the form "day month year".
No default.}

\item{keep_all}{a boolean; default FALSE.  If true, won't filter data based
on a_date; it will also change a_date to today's date.}
}
\value{
A data frame with one row per read attempt.
}
\description{
Shows book completion stats on a date specified by user.
}
\details{
Supplied with a dataframe of book notes and a date of interest, this
will return the dataframe supplied to it with additional columns to show how
long a read attempt has taken (and whether it has been completed or just set
with an end date equal to the current system date).
}
