% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reading.R
\name{check_book_progress}
\alias{check_book_progress}
\title{Show Book Progress on a Particular Day}
\usage{
check_book_progress(my_data = NULL, a_date = NULL)
}
\arguments{
\item{my_data}{a dataframe of book notes.  Must include start and finish
dates for reading attempts.  Default NULL.}

\item{a_date}{a date or character object of the form "day month year".
default NULL.}
}
\value{
The same dataframe, with the following additional columns:
  \describe{
    \item{unfinished_1}{boolean; whether read_1 finished the book}
    \item{unfinished_2}{boolean; whether read_2 finished the book}
    \item{relevant}{used to filter books}
    \item{duration_1}{integer; how many days read_1 took;
      equals today if unfinished_1 TRUE}
    \item{duration_2}{integer; how many days read_1 took;
      equals today if unfinished_2 TRUE}
  }
}
\description{
Shows book completion stats on a date specified by user.
}
\details{
Supplied with a dataframe of book notes and a date of interest, this
function will first filter out books that had unparsable finish dates.  It
will return the dataframe supplied to it with additional columns to show how
long a read attempt has taken (and whether it has been completed or just set
with an end date equal to the current system date).
}
