% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reading.R
\name{read_book_notes}
\alias{read_book_notes}
\title{Read in My Book Notes}
\usage{
read_book_notes(source_file = NULL, remove_mess = TRUE)
}
\arguments{
\item{source_file}{path to the plain-text file to pull book notes from.
Will only work if used on a file with the same layout as mine.
Default NULL.}

\item{remove_mess}{a boolean that determines whether the initial
messy-looking variables are removed so that the final data frame produced
is properly tidied.  Default TRUE.}
}
\value{
a data frame with bibliographic details of books as separate
  columns.  Also includes start and end dates of reading each book and
  keywords I've noted for each.
}
\description{
Reads in a plain text file of notes on books and returns a tidy data frame.
}
\details{
Designed for my own format of notes on books, including when I read them.
Won't work on anything that doesn't match that formatting.  Currently does
the bare minimum of creating sensible variables.  Doesn't prettify anything
yet.
}
\keyword{data}
\keyword{idiosyncratic,}
\keyword{munging,}
