% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reading.R
\name{read_book_notes}
\alias{read_book_notes}
\title{Read in My Book Notes}
\usage{
read_book_notes(
  source_file = NULL,
  remove_mess = TRUE,
  keep_ugly_dates = FALSE,
  diagnose_speed = FALSE
)
}
\arguments{
\item{source_file}{path to the plain-text file to pull book notes from.
Will only work if used on a file with the same layout as mine.
Default NULL.}

\item{remove_mess}{a boolean that determines whether the initial
messy-looking variables are removed so that the final data frame produced
is properly tidied.  Default TRUE.}

\item{keep_ugly_dates}{a boolean; if TRUE, the data frame returned will 
include the character version of the dates as well as the cleaned-up
date class columns.  Default FALSE.}

\item{diagnose_speed}{a boolean; if TRUE, returns a self-diagnosis instead
of the notes on books.  Default FALSE.}
}
\value{
a data frame with bibliographic details of books as separate
  columns.  Also includes start and end dates of reading each book and
  keywords I've noted for each.  If asked to, will return a diagnosis of the
  function speed instead of the data frame.
}
\description{
Reads in a plain text file of notes on books and returns a tidy data frame.
}
\details{
Designed for my own format of notes on books, including when I read them.
Won't work on anything that doesn't match that formatting.  Creates
tolerably pretty variables.  If asked to, will return a report of how long
each step of the function took instead of returning the notes.
}
\keyword{data}
\keyword{idiosyncratic}
\keyword{munging,}
